+++ crates/try_v2_traits/src/transform.rs
@@ -1,13 +1,16 @@
 use std::ops::{ControlFlow, FromResidual, Residual, Try};
 
+impl<T> Transform<T> for Option<T> {}
+impl<T, E> Transform<T> for Result<T, E> {}
+impl<B, C> Transform<C> for ControlFlow<B, C> {}
+
 /// Methods for transforming TryTypes. Inspired by the methods provided on `Option` & `Result`
 ///
 /// ## Note
 ///
-/// - Methods which act on the contained value are only available for the *Output* case. TryTypes
-///   are recommended to directly implement equivalent methods for *Residual* cases with suitable
-///   naming. E.g. we provide a `.map()` but not a `.map_err()` equivalent as multiple such
-///   methods may be needed and no standardised naming makes sense.
+/// - TryTypes are recommended to directly implement specific `map_xxx` methods for *Residual* cases
+///   with suitable where this makes sense. `map_residual()` provides for the cases where generic
+///   naming is acceptable, and/or for implementations generic over `Try + Transform`
 /// - Methods which act on the contained value will extract a value of type `Output` and return the
 ///   *canonical TryType* for the new Output. This is identifiable by the generic type `X` in the
 ///   method signature. This is usually the expected behaviour but can lead to a different value
@@ -19,7 +22,7 @@ use std::ops::{ControlFlow, FromResidual, Residual, Try};
 ///     - `U` the other `Output` type
 ///     - `F` a function/closure passed as a parameter
 ///     - `G` the return type of `F`
-///     - `R` *never used* to avoid confusion with "Residual".
+///     - `R` *never used* to avoid confusion with "Return" vs. "Residual".
 pub trait Transform<T>
 where
     Self: Try<Output = T> + Sized,
@@ -88,6 +91,58 @@ where
         match self.branch() {
             ControlFlow::Continue(val) => f(val),
             ControlFlow::Break(_) => default(),
+        }
+    }
+
+    /// Applies a function to the residual, leaving output values untouched, then returns the
+    /// canonical TryType for the residual returned by the function (and Output `T`).
+    ///
+    /// # Note:
+    /// - `f` takes the Residual and returns a Residual, not any values potentially wrapped by the
+    ///   Residual. This is different from the specialised `Result::map_err()` which works on the
+    ///   wrapped error value.
+    ///
+    /// ```
+    /// # use try_v2_traits::Transform;
+    /// let err_5: Result<String, i32> = Err(5);
+    ///
+    /// // Err(e).map_err() operates directly on e
+    /// let stdlib = err_5.clone().map_err(|e| format!("{e}"));
+    ///
+    /// // Err(e).map_residual() operates on Err(e)
+    /// // With only one residual variant direct destructuring is possible
+    /// let custom = err_5.map_residual(|Err(e)| Err(format!("{e}")));
+    ///
+    /// assert_eq!(stdlib, custom);
+    /// ```
+    ///
+    /// - This allows for interconversion between TryTypes if `f` returns a different residual:
+    ///
+    /// ```
+    /// # use try_v2_traits::Transform;
+    /// # use std::ops::ControlFlow;
+    /// let err_5: Result<String, i32> = Err(5);
+    /// let custom = err_5.map_residual(|Err(e)| ControlFlow::Break(e));
+    /// assert_eq!(custom, ControlFlow::Break(5))
+    /// ```
+    ///
+    /// ```
+    /// # use try_v2_traits::Transform;
+    /// // This is effectively the equivalent of Option::ok_or_else()
+    /// let none: Option<i32> = None;
+    /// let stdlib = none.ok_or_else(|| 5);
+    /// let custom = none.map_residual(|_| Err(5));
+    /// assert_eq!(stdlib, custom);
+    /// ```
+    fn map_residual<F, X, G>(self, f: F) -> X
+    where
+        F: FnOnce(Self::Residual) -> G,
+        G: Residual<T, TryType = X>,
+        X: Try<Output = T> + FromResidual<G>,
+    {
+        match self.branch() {
+            ControlFlow::Continue(val) => X::from_output(val),
+            ControlFlow::Break(residual) => X::from_residual(f(residual)),
         }
     }
 
@@ -220,9 +275,6 @@ where
 mod tests {
     use super::*;
 
-    impl<T> Transform<T> for Option<T> {}
-    impl<T, E> Transform<T> for Result<T, E> {}
-
     mod boolean {
         use super::*;
 
@@ -342,6 +394,21 @@ mod tests {
             let stdlib = some_5.map_or_else(|| 1 + 1, |x| x + 1);
             let custom = Transform::map_or_else(some_5, || 1 + 1, |x| x + 1);
             assert_eq!(stdlib, custom);
+        }
+
+        #[test]
+        fn map_residual() {
+            let err_5: Result<String, i32> = Err(5);
+            let stdlib = err_5.clone().map_err(|e| format!("{e}"));
+            let custom = err_5.map_residual(|Err(e)| Err(format!("{e}")));
+            assert_eq!(stdlib, custom);
+        }
+
+        #[test]
+        fn map_residual_conversion() {
+            let err_5: Result<String, i32> = Err(5);
+            let custom = err_5.map_residual(|Err(e)| ControlFlow::Break(e));
+            assert_eq!(custom, ControlFlow::Break(5))
         }
     }